requirements.txt
flask
flask-cors
requests
python-docx
python-pptx
